backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
google-auth
requests
